Exception handling

+++ src/main/java/com/devmarrima/dscommerce_list/services/ProductService.java
@@ -1,13 +1,19 @@
 package com.devmarrima.dscommerce_list.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import com.devmarrima.dscommerce_list.dto.ProductDTO;
 import com.devmarrima.dscommerce_list.entities.Product;
 import com.devmarrima.dscommerce_list.repositories.ProductRepository;
+import com.devmarrima.dscommerce_list.services.exceptions.DataBaseException;
+import com.devmarrima.dscommerce_list.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -17,7 +23,8 @@ public class ProductService {
 
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 	}
 
@@ -37,17 +44,33 @@ public class ProductService {
 
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado!");
+		}
+
 	}
 
-	@Transactional
-	public void delete (Long id) {
-		repository.deleteById(id);
-	}
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id) {
+		if(!repository.existsById(id)){
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+			try{
+				repository.deleteById(id);
+			}
+			catch(DataIntegrityViolationException e){
+				throw new DataBaseException("Violação de restrição de integridade referencial!");
+			}
+		}
+
 	
+
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
 		entity.setName(dto.getName());
 		entity.setDescription(dto.getDescription());
